+++ src/components/buttons/createTaskListButton/createTaskListButton.tsx
@@ -1,13 +1,13 @@
 import {faPlus} from '@fortawesome/free-solid-svg-icons';
 import {FontAwesomeIcon} from '@fortawesome/react-native-fontawesome';
+import {ReturnComponentType} from 'commonTypes/returnComponentType';
 import React, {useState} from 'react';
 import {useDispatch} from 'react-redux';
 import {TaskListType} from 'store/reducers/taskListReducer/types';
-import {ReturnComponentType} from 'types/common/returnComponentType';
-import {iconSizeLarge} from '../../../../../constants/constants';
-import {addNewTaskList} from '../../../../../store/actions/tasksActions/taskListActions';
-import {Input} from '../../../../common/input/Input';
-import {ModalIcon} from '../../../../common/modals/ModalIcon';
+import {iconSizeLarge} from '../../../constants/constants';
+import {addNewTaskList} from '../../../store/actions/tasksActions/taskListActions';
+import {Input} from '../../common/input/Input';
+import {ModalIcon} from '../../common/modals/ModalIcon';
 
 export const CreateTaskListButton = (): ReturnComponentType => {
   const dispatch = useDispatch();
