+++ src/components/buttons/deleteTaskButton/DeleteTaskButton.tsx
@@ -3,14 +3,14 @@ import {FontAwesomeIcon} from '@fortawesome/react-native-fontawesome';
 import React from 'react';
 import {Text} from 'react-native';
 import {useDispatch} from 'react-redux';
-import {iconSizeSmall} from '../../../../../constants/constants';
+import {ReturnComponentType} from '../../../commonTypes/returnComponentType';
+import {iconSizeSmall} from '../../../constants/constants';
 import {
   deleteTask,
   deleteTaskListFromScreen,
   deleteTaskListFull,
-} from '../../../../../store/actions/tasksActions/taskListActions';
-import {ReturnComponentType} from '../../../../../types/common/returnComponentType';
-import {ModalIcon} from '../../../../common/modals/ModalIcon';
+} from '../../../store/actions/tasksActions/taskListActions';
+import {ModalIcon} from '../../common/modals/ModalIcon';
 import {styles} from './styles';
 import {DeleteTaskButtonPropsType} from './types';
