+++ src/components/buttons/deleteTaskButton/styles.ts
