+++ src/components/buttons/deleteTaskListButton/DeleteTaskListButton.tsx
@@ -3,13 +3,13 @@ import {FontAwesomeIcon} from '@fortawesome/react-native-fontawesome';
 import React from 'react';
 import {Text} from 'react-native';
 import {useDispatch} from 'react-redux';
-import {iconSizeSmall} from '../../../../../constants/constants';
+import {ReturnComponentType} from '../../../commonTypes/returnComponentType';
+import {iconSizeSmall} from '../../../constants/constants';
 import {
   deleteTaskListFromScreen,
   deleteTaskListFull,
-} from '../../../../../store/actions/tasksActions/taskListActions';
-import {ReturnComponentType} from '../../../../../types/common/returnComponentType';
-import {ModalIcon} from '../../../../common/modals/ModalIcon';
+} from '../../../store/actions/tasksActions/taskListActions';
+import {ModalIcon} from '../../common/modals/ModalIcon';
 import {styles} from './styles';
 import {DeleteTaskListButtonPropsType} from './types';
