+++ src/components/buttons/editTaskListTitleButton/EditTaskListTitleButton.tsx
@@ -2,11 +2,11 @@ import {faPen} from '@fortawesome/free-solid-svg-icons';
 import {FontAwesomeIcon} from '@fortawesome/react-native-fontawesome';
 import React, {useState} from 'react';
 import {useDispatch} from 'react-redux';
-import {iconSizeSmall} from '../../../../../constants/constants';
-import {setEditedTaskListTitle} from '../../../../../store/actions/tasksActions/taskListActions';
-import {ReturnComponentType} from '../../../../../types/common/returnComponentType';
-import {Input} from '../../../../common/input/Input';
-import {ModalIcon} from '../../../../common/modals/ModalIcon';
+import {ReturnComponentType} from '../../../commonTypes/returnComponentType';
+import {iconSizeSmall} from '../../../constants/constants';
+import {setEditedTaskListTitle} from '../../../store/actions/tasksActions/taskListActions';
+import {Input} from '../../common/input/Input';
+import {ModalIcon} from '../../common/modals/ModalIcon';
 import {EditTaskListTitleButtonPropsType} from './types';
 
 export const EditTaskListTitleButton = ({
