+++ src/components/common/taskList/TaskList.tsx
@@ -1,10 +1,10 @@
+import {ReturnComponentType} from 'commonTypes/returnComponentType';
 import React, {FC} from 'react';
 import {FlatList, ListRenderItem, Text, View} from 'react-native';
-import {ReturnComponentType} from 'types/common/returnComponentType';
 import {TaskType} from '../../../store/reducers/taskListReducer/types';
-import {CreateTaskButton} from '../../navigation/tasksNavigation/buttons/createTaskButton/CreateTaskButton';
-import {DeleteTaskListButton} from '../../navigation/tasksNavigation/buttons/deleteTaskListButton/DeleteTaskListButton';
-import {EditTaskListTitleButton} from '../../navigation/tasksNavigation/buttons/editTaskListTitleButton/EditTaskListTitleButton';
+import {CreateTaskButton} from '../../buttons/createTaskButton/CreateTaskButton';
+import {DeleteTaskListButton} from '../../buttons/deleteTaskListButton/DeleteTaskListButton';
+import {EditTaskListTitleButton} from '../../buttons/editTaskListTitleButton/EditTaskListTitleButton';
 import {Task} from '../task/Task';
 import {styles} from './styles';
 import {TaskListPropsType} from './types';
