+++ src/components/screens/signInScreen/SignInScreen.tsx
@@ -1,6 +1,6 @@
+import {ReturnComponentType} from 'commonTypes/returnComponentType';
 import React from 'react';
 import {Text, View} from 'react-native';
-import {ReturnComponentType} from 'types/common/returnComponentType';
 import {styles} from './styles';
 
 export const SignInScreen = (): ReturnComponentType => {
